Use better enum for operations
Previously it was necessary to check what is the character we observe
in order to determine which operation are we currently parsing.
This is now implemented in the following way:

1. `enum Operation` is now implemented and contains additional option
    Operand(char) which contains the actual character
2. `enum Operation` has functions `from_operand` and `order`
    - `from_operand(&self)` returns the Operation
    - `order(&self)` returns the order of operation

Also, `enum OperationOrder` is introduced to better reason about the
order of the operation itself. i.e. `Operation::First` get's calculated
before `Operation::Second` etc.